remove google analytics
As of now, we are only choosing to embed our code into another page to
enhance the user's experience.

However, this means it is dubious to try to inject our GA script.

I will attempt to use this again in the future if I decide to move logic
to a separate window or popup.